Refactor some codes for further implementation.
